fixed the algorithm to accurately sort numbers. find_index was off by 1
find_steps to the top was off by 1 as well

function m was rotating the stack in the wrong order, was missing a condition

+++ ft_strtol.c
@@ -33,7 +33,7 @@ long	handle_overflow(unsigned long result, int digit, int sign)
 	int				cutlimit;
 
 	if (sign == -1)
-		cutoff = -(unsigned int)INT_MIN;
+		cutoff = -(long)INT_MIN;
 	else
 		cutoff = INT_MAX;
 	cutlimit = cutoff % 10;
